OM-26182: AO Settings: Creating a cache for entity settings that all Action objects refer to.
We were previously passing a Map<> to each action, but that made it difficult to update setting definitions for re-recommended actions. There were two ways to go forward:
1) Have a method to override the Map<> in the action.
2) Move the Map<> into a helper class, and have the action refer to the helper class.

This change implements option 2. There is now an EntitySettingsCache in the Action Orchestrator which contains settings for all entities involved in READY actions. The cache is updated every time we process a new ActionPlan from the market. We calculate the set of all involved entity IDs when processing the action plan (only get IDs for READY state actions), and make a single call to the group component to retrieve the settings. There is no retry logic, so if the call fails there will be no settings until the next broadcast.

Testing Done:
- Full build.
- Integration testing - rebroadcasting topology with modified setting policies leads to all action modes being change.

Reviewed By: mneuburger
https://rbcommons.com/s/VMTurbo/r/19367/