OM-27354: XL Settings: Supporting multiple entity types per setting.
Our current approach is to emulate the settings as they are returned by legacy, with a separate setting for every entity type (e.g. suspendVM, suspendPM). This is a big headache when dealing with settings, because usually we are dealing with a generic entity and a generic setting, and we end up having to do prefix searches.

This change adds support for multiple entity types per setting.

The main change is in the API component, where we need to provide SINGLE entity types in responses.

For setting specs, the SettingApiDTO (contained in the SettingManagerApiDTO) needs an entity type:
- When retrieving setting specs, an entity type is usually provided (always provided from the UI). If it's provided, we just set the SettingApiDTO entity type to the provided one (assuming the provided entity type is one of the supported types by the spec).
- If an entity type is not provided, and we need to return information about a setting spec that supports multiple
entities, we just pick the first one. This isn't quite right, but there's nothing else we can do (other than not set an entity type at all)

For setting policies, the SettingPolicyApiDTO input doesn't have an entity type. We used to infer the entity type from the SettingSpec, but we can't do that anymore. So instead:
- When creating a new setting policy we look at the scopes (all user-created SPs have scopes), and get the entity type from there.
- When editing a setting policy we look at the entity type of the existing setting policy.

I also added the "suspend" setting (without support in the rest of the system yet) because hosts don't support "move."

Testing Done:
- Full build
- CRUD operations on Suspend setting policies (because those apply to VMs and PMs)

Reviewed at https://rbcommons.com/s/VMTurbo/r/19667/