OM-27357: XL History: Fixing intermittently failing test.
The LiveTopologyListenerTest had a race condition, where depending on the speed of the underlying system the two topologies may not have been interleaved.

For example:
- topology1 and topology2 get scheduled for processing in the executor
- Latch unlocked (allowing processing to complete)
- Thread for topology1 runs
- Thread for topology1 completes
- Thread for topology2 runs
- Thread for topology2 completes
- TEST FAILURE!

The solution is to keep the latch locked (blocking all processing) for topology1, and run topology2 processing once the thread for topology1 is blocked on the latch. The new workflow is:
- topology1 gets scheduled for processing
- Wait until it hits the latch (using Mockito.verify)
- Process topology2 on the main thread
- Verify that topology2 got dropped
- Latch unlocked
- Wait for topology1 thread to finish
- TEST SUCCESS!

Testing Done:
- Ran the test. It passed before and now on my system, but by inspection it should run better on Jenkins now :)

Reviewed at https://rbcommons.com/s/VMTurbo/r/19685/