OM-27320 RB:19916 Use entity types to determine move settings
Store entity types within entity settings cache and consult them to determine if a move action is specifically a storage move.
Since the entity types are stored within the repository component which may not be in sync with the action orchestrator,
the fetch happens asynchronously with a configurable retry.

Future work: support unlimited retry, convert entity type retrieval to GRPC service, differentiate Move/Storage Move action types so that repository query is not needed