OM-27391: Add function to compute scoped topology to the Analysis package RB-17778: Roman, Patrick
Use the existing functions to convert the topology to Trader Objects, and then create an Economy.

Query the economy to determine the "upwards" sells-to recursively to the "top", and then follow the "downwards" may-sell-to relationships recursively, and collect elements of the scoped topology "both upwards and downwards".

The new method will be called in OM-27566

This is covered by the WIKI page: https://vmturbo.atlassian.net/wiki/spaces/Home/pages/277774451/Constructing+Scoped+Topology

Testing:
- added unit tests for scoped topologies
- all unit tests pass
- tested on live system running a plan