OM-27515: Capacity and usage for clusters is showing data for only one PM RB-20080: Karthik
The Capacity was not being accumulated in the loop over service entities.

Also simplified the avg/total calculation code.

Removed duplicate code from a prior test: testGetStatsCounts

Added unit test for the capacity, as well as the other avereaged values.
Tested in running component and the Capacity reflects all the cluster/group members