XL: Setting CPU and MEM effective capacity to HA * max desired state for cluster headroom plans.
When running headroom calculation, we want the CPU and MEM capacity of hosts in the cluster to reflect the desired state as well as HA reservation. This is because we want the headroom projection to accurately represent the state that the market would recommend - which would take desired state into account.

The analysis for headroom itself does NOT take desired state into account, because desired state controls provisions and moves - neither of which are allowed in a headroom analysis. So we factor the desired state into the effective capacity available to each host.

Changes:
- Injecting TopologyInfo into the EntitySettingsApplicator, so that the type of topology we're analyzing can affect the way we interpret settings.
- Creating the list of applicators dynamically instead of statically, so that TopologyInfo can be passed to the one stage that needs it. The alternative was to make the "apply()" method take the TopologyInfo as well, but I thought that was overkill for now.

Testing Done:
- Unit tests to confirm the values get set properly.

Bugs closed: OM-28046

Reviewed at https://rbcommons.com/s/VMTurbo/r/20114/