XL API: Return a mutable list from ActionSpecMapper, so that caller can modify it without making a copy.
From an interface point of view, it doesn't really make sense to return an immutable list - the caller should be able to do whatever they want with the list.

From a practical standpoint, the "default" pagination in com.vmturbo.api uses an in-place sort to order the action list, and the in-place sort fails if the markets service returns an immutable list.

Testing Done:
- Tests pass.
- Getting a page of results via the REST API works.

Bugs closed: OM-29540

Reviewed at https://rbcommons.com/s/VMTurbo/r/20816/