OM-28938 : Create Data Retention SettingSpec and implement "settings/persistencemanager" rest api RB:20610 OM-28312 : Add ability to set history/stats data retention values
retention_policies table in vmtdb exists only in XL. But the jooq classes are generated
by the reporting component in legacy opsmgr. Inorder for jooq classes to be generated
for retention_policies table, a new sql definiton is addedd (V72_3__AddRetentionPoliciesTable.sql)
to to legacy opsmgr. This table will be empty and not used in legacy opsmgr.

* Add global setting specs for stats, audit, plan and reporting data retention settings
* Rename SettingPolicySetting -> EntitySettingSpecs
* Move EntitySettingSpecs, GlobalSettingSpecs and related code to common-protobuf

Review : https://rbcommons.com/s/VMTurbo/r/20610/