OM-29226: Add pre-check in component startup to require remote configuration before starting RB-20920: Patrick, Karthik
Add pre-check in component startup to require remote configuration before starting

The main goal is to make ClusterMgr the sole source of configuration properties instead of going directly to Consul, which leads to a race condition with ClusterMgr initializing Consul.

Remove the @EnableAutoConfiguration annotation, and insert a call to the ClusterMgr API to fetch configuration properties before Spring Instantiation. This changes the "main()" for each component.

The config properties fetch will wait until the clustermgr responds, and then the Spring Boot Initialization is called.