OM-29989: -address in use- error prevents consul from starting after reboot RB-21077: Patrick
Change the fixed IP address for consul to a higher number to avoid the IP's automatically assigned by Docker on restart.

Note: consul needs a static IP so it can know "this is me" the next time it comes up" - otherwise it tries to negotiate with a consul at the previous IP, which will never answer.

Also...in an unrelated change:  added the multi-threaded option to the xz comporession command to speed up ISO generation. In a test with level 0 compression the time went down from 6:50 minues for one of the images to 1:45; and we compress at level 9. I can definitely see that all processors are being used on my Mac during the xz step, much more than before this change.