OM-29279 RB:21193 Fix plan supply chain fetch failure
The UI was attempting to fetch the supply chain for a plan after running by passing
the plan ID as a "seed UUID" and the live topology context ID to the supply chain
fetcher. It should have been passing the plan ID as the topology context ID, and did
not need any seed UUIDs. In the long run a different API call for plan vs market supply
chain retrieval might be good to have, but in the short term the service attempts to
determine if the call is meant for a plan topology or a live topology and formats the request accordingly.