OM-30242: Speed up build of mediation components RB-21248: Karthik
Specify an entire folder be included for probe-jars, instead of one file at a time. This requires only a single docker image layer for the folder copy, instead of an individual layer for each file.

Note that there seems to be a problem with the docker-maven-plugin where the first <resource> entry must have an <includes> or it doesn't work at all; but subsequent <resource> entries do not need the <includes> to work.