OM-29719 RB:21573 Add Virtual Datacenter CPU Allocation capacity post-stitching operation
New post-stitching operation for setting the CPU Allocation capacity of Virtual Datacenters for which it isn't already set.

If a Virtual Datacenter has another Virtual Datacenter as a provider, the provider's CPU Allocation capacity is used. Otherwise, the capacity is set to the sum of the CPU Allocation capacities of all of its Physical Machine providers.