OM-30249: Implement widgetsets rpc in auth component RB-21332: David
Implement widgetsets rpc in auth component - a wrapper around the persistence layer to be done in OM-30251.

Added the WidgetsetRpcService to the Auth Component, including an Interface for the persistence store.

Added a new project auth-test for a test utility JwtContextUtil, which will be needed by any component to access the JWT for identity and authorization