XL: Create and use a common RestTemplate with properly ordered message converters.
We are having a problem where the default RestTemplate puts the XML converter before the JSON converter. This makes it necessary to explicitly specify the request content type in order to use JSON serialization.

The (quick fix) solution is to have a common RestTemplate for every component in the system that places the JSON serializer (using GSON) above the XML one in the ordering.

The more long-term fix is to complete our gRPC migration!

Testing Done:
- Full build.
- Started up XL. Calls that used to fail (e.g. init admin) no longer fail.

Bugs closed: OM-30996

Reviewed at https://rbcommons.com/s/VMTurbo/r/21633/