OM-30392: Fix db query time range for latest table queries RB-21788: Gary
For the "before" panel of the Optimized Improvements, the UI requests stats from a single point in time, i.e. startTime = endTime based on the current time-of-day and the UI time range. Of course time samples will not fall at a time equal to the startTime

move the time-window for querying _latest table to external configuration.

For the queries to the "_by_hour", "_by_day", and "_by_month" tables the startTime is mapped to the beginning of the hour/day/month.

For queries to the _latest table we need to create a time range around the UI time. In this review we use a 10-minute window to match the broadcast time.