OM-30356 RB:21836 User settings were not overriding default settings
User settings were being added as duplicates and not overriding default settings. As a result,
both settings were still present, and the more restrictive one was chosen even if it was default
and should not have been there. The fix was to make sure that duplicate default settings were
not added by checking against the spec names of the user settings already there.