OM-30288 : Display the correct host count when a Cluster is selected from the UI. RB:21798
VCenter cluster has VDC as one of its members along with the PhysicalMachines.
That's why in the UI, the count for hosts was alwasy showing NumOfPMs+1. Checked
with Dimtri. He said that filtering the VirtualDataCenter from the Cluster
member list is fine.

Generalized this to not allow clusters/groups to have entities of different
types. When creating the  member list for groups/clusters, we filter out any
entity whose type doesn't match the entityType of the group.

https://rbcommons.com/s/VMTurbo/r/21798/