OM-31285: XL API: Do not pull ServiceConfig into dispatcher servlet.
The @Import of ServiceConfig in the DispatcherControllerConfig caused a re-creation of all the beans in ServiceConfig (and it's imports) in the Dispatcher Servlet's application context. This caused some bugs (e.g. Kafka client being initialized more than once, plan deletion not working properly).

Fixing this by removing the @Import and adding some comments, and auto-wiring the License Service directly.

Testing Done:
- Verified that there are no double log entries upon component startup.
- Testing plan deletion - it works now, whereas it didn't before.

Reviewed By: mark.laff
https://rbcommons.com/s/VMTurbo/r/21881/