OM-29783 RB:21955 Added MustNotRunTogetherPolicy + changed the behavior of MustRunTogether, in order to be similar to the MustNot one. Before this change, the MustNotRunTogether policy was modeled using the AtMostN policy, which forces you to specify a provider group for it.
This is semantically different from just a MustNotRunTogether policy, because in this case you don't want (or you don't know) which ones are the providers that you want to use as limit.
Ideally you only know that you don't want ever those 2 entities to run together.
Because of this, I changes the MustRunTogether to only take a group of entities (the consumers), and on which type of provider you want to enforce the running together (run on same host, or same storage, ...).

The same idea is applied to MustNotRunTogether: only 1 group + the provider type.

The providers that needs to sell the segmentation comm are calculated at policy application time (instead of at policy creation time).