OM-30159: API: Handle special case of retrieving supply chain of a group, restricted to the entity type of the group.
There is a common (i.e. done it at least two places) pattern of asking for the members of the supply chain scoped to a group, restricted by entity type to the entities in a group. For example: get the supply chain (or supply chain stats) scoped to group 123 (of PhysicalMachines), with entityType restricted to PhysicalMachine. The way we handle this case right now is to get the single-source supply chain of each entity in the group, and merge them. This is incredibly inefficient when we get to thousands of entities in a group (e.g. when clicking around on the global supply chain).

This review adds special handling for this use case to the supply chain fetcher in the API. If the request is for a single group, and the entity type asked for in the request is the same as the entity type of the group, we bypass the repository's supply chain API and "fake" a SupplyChainNode composed of the members of the group. The rest of the supply-chain related logic remains the same - as if the supply chain API returned the node.

Testing Done:
- Ran build + unit tests.
- Started up updated API component. In XL with 50k entities (5k hosts), the API calls to get the supply chain for 5k hosts complete instead of timing out! The health widget renders in under-1 minute the first time, and faster the following times, instead of never rendering.

Reviewed at https://rbcommons.com/s/VMTurbo/r/21837/