OM-24696: Host-based charts in Details tab are empty when scoped to DataCenter RB-21941: Patrick, Roman
In 6.x there is an expansion step in gathering the list of SE's to calculate stats for, where each DataCenter SE is replaced by the PMs in that DC.

We implement this using the supplychain function as is done in 6.x.

Note that this should handle individual DCs as well as groups of DCs.