OM-32006: Commodity list returned not filtered by relatedEntity parameter RB-21992: David
The UX is sending a request for /stats/Market but with a 'relatedEntity' parameter.  This means that stats should be limited to the given entity type.  We added a method to use that entity type to determine which stats table to query, and then use the existing query generation and result return mechanism.

(cherry picked from commit f62a6a0)