OM-29172 RB:22023 Group resolution no longer writes extraneous errors to the log
Previously when attempting group resolution on an id that did not correspond to a group, the group
service would log an error even if returning an empty result was a supported part of the workflow
e.g. with the GroupExpander. Now if the expectPresent flag is false (as opposed to true, which is its
default) then an empty result from attempting to resolve a non-group entity will not log an error.