OM-30349: Exception in HistorydbIO.persistEntities due to duplicate entry for key PRIMARY. RB-22160: Pavel
In the previous change to optimize entity table writes by batching there was no provision for an entity to change displayname or type.

It turns out displayName will may change under user control, but in this case there were two targets in which the same entity (shared Storage) was given different names.

This particular storage appears as QSX:NFSShare in dc12 and QS1:NFSShare in dc13.

So...depending in which target was discovered last, the displayName would be changed, and the LiveStatsWriter tried to insert a new record instead of changing the previous record.

In this change we fetch the DB rows using jooq, and store the modified rows in case there is a change as in the case describe above.

There is no change for persisting new entities.

(cherry picked from commit b9c2588)