OM-32470: Clustermgr swagger is broken and cannot be accessed. RB-22215: Patrick
ClusterMgr Swagger used to depend on the SpringFox  @EnableSwagger2 annotation, which does a component scan (== slow).
When we switched the base VMTComponents to use pre-generated swagger.json, ClusterMgr (which is not a VMTComponent) got left behind.
I tried to convert ClusterMgr to a VMTComponent but ran into JSON conversion problems which would have required re-engineering too much to fix.
In the meanwhile I copied the pom.xml swagger-maven-plugin generation and WebMVC redirection from the base component pom.xml to clustermgr/pom.xml.
In the future we will be redesigning ClusterMgr in the context of Kubernetes, and that will be the time to resolve these overlaps.

(cherry picked from commit 01703f5)