OM-31827 : Remove entityTypeMap from Action Orchestrator. RB:22204
Market has the entire topology. Hence it already knows about the entityTypes for
the entities specified in the Actions. Market will set the entityType along
with the entityIds before sending it to Action Orchestrator(AO).
This eliminates the need for querying Repository from AO to fetch just the
entity types for the given entityIds. Quering the repo is a heavy operation.
So this change should also help improve the performance of AO.

RB: https://rbcommons.com/s/VMTurbo/r/22204/