OM-32630: XL: Implement new APIs to get per-entity stats from plans.
We added new APIs to the IMarketsService to retrieve per-entity stats for plans, and group in plans. This change adds support for those API calls in XL.

Adding an "EntitiesFilter" containing a list of ids to the Repository's RPC to get plan entity stats. In the repository, only apply the stats filters to entities in the projected topology that contain the requested IDs.

Still TODO (not in this review):
- We do not store risk index for projected topologies, so the sorting by post-plan risk index in the optimized improvements widget won't work.
- We do not currently the "relatedType" argument in the per-entity stats. It's unclear what exactly the impact of that is on the optimized improvements widget; needs more testing.

Testing Done:
- Tested running a plan with the updated optimized improvements widget (in review), and saw post-plan entity stats returned.
- Build

Bugs closed: OM-32630

Reviewed at https://rbcommons.com/s/VMTurbo/r/22236/