OM-32544: XL: Increasing the max message size of group clients to deal with large "entity settings" responses. Adding GroupPerformanceTest.
In a large topology, the GetEntitiesSettingsResponse message was larger than the maximum default message size for gRPC. This caused errors filling the EntitySettingCache in the action orchestrator.

This change increases the message size for group clients. I opened another bug (OM-32762) to track the effort of reducing the message size (by refactoring or chunking), but it's unnecessary for GA because the call completes in <1s (in the newly added GroupPerformanceTest).

Also, adding the GroupPerformanceTest that I used to reproduce the issue. Right now it only measures performance of the EntitySettingStore.

Testing Done:
- Full build.
- GroupPerformanceTest failed without new max message size, succeeds now.

Bugs closed: OM-32544

Reviewed at https://rbcommons.com/s/VMTurbo/r/22307/