OM-32745: Create new History DB Baseline RB-22369: Greg, Roman, Karthik branch: develop
1) run XL 7.0.0 DB build on empty db
2) sqldump to .sql file 'v1.0__baseline.sql' file in a new project: com.vmturbo.history.schema
3) remove all the old XL migration files
4) remove all references to opsmgr project com.vmturbo.reports; copy necessary classes to XL
5) change the flyway migration folder property to point to this new v1.0__baseline.sql

In other words, we have completely broken away from dependencies on com.vmturbo.reports from classic.

This exposed a difference between the Jooq classes for classic and the DB tables for XL:  the _by_week and _by_month tables in Classic are 'date' and in XL are 'timestamp'. Some of the code in XL needed to be updated to accurately capture the datatype being stored/returned.

(cherry picked from commit 3b4cc67)