OM-32969: The flyway/jooq setup should establish the privileged DB user (vmtplatform) if necessary RB-22432: Greg develop
The sql for several modules was dumped as 'created by user vmtplatform'. Thus, to be loaded, that userid must be defined. We do that with an additional migration file, with number "0.1" so that it will be run first.

We don't, however, want to distribute the db password with the file, so this userid-definition migration file is located in the test/resources folder.

I realize that this small migration file is the same between the two projects, but that may ony be a happy accident. I feel it is important to keep the two files separate, and not create a cross-package reference.

(cherry picked from commit e5731a4)