OM-30390: optimized improvements plan before chart empty RB-22489: Patrick, Roman develop
If startTime=endTime we were searching a window around startTime for stats, where the width of the window was equal to the broadcast interval (was 10 mins).

So we were looking 5 mins before the startTime and 5 mins after the start time for the "before" stats from the live market.

If a plan was run > 5 mins after the most recent broadcast (e.g. 8 minutes afterwards), the request for stats for "plan time" would only look back to "plan time - 5 minutes", i.e. not far enough to see the most recent broadcast. Further, we were ajusting the "endTime" to 5 minutes in the future, which wouldn't help find any values.

Instead, if the startTime == endTime, we adjust the startTime backwards the full window width (increased to 15 mins in case processing large topologies is slow) and don't modify the endTime. In this example, we will look at "plan time - 15 minutes" and will find the most recent live stats (at plan time - 8 minutes).

(cherry picked from commit d38280e)