OM-29189, OM-32469 RB:22576 Avoid prematurely filling in commodity values by postponing entity validation
Previously if an entity was discovered with an illegal or unset commodity value, that value would be
replaced during discovery validation. As a result the relevant post-stitching operation later on would
not be applied, because the value was already replaced with an incorrect one. This change moves
validation and replacement to later in the topology pipeline, after the post-stitching phase, so that any
values that must be calculated are already present. This makes sure there are no false positives during
validation and no false negatives during post-stitching.

Also, rearrange post-stitching operations so that all storages have their access capacity set by the
appropriate operation depending on whether a provider sets it for them.