OM-33181: On restart/update the "old" version of the containers start running while the update is in progress. RB-22541: Karthik, Billy, Greg, Patrick
The curent behavior is for the API component to displayed within a minute of restarting the VM; the user may log in, but it will be the "old" (pre-update) version of the UX, and as the update proceeds various internal containers will be stopped / restarted as they are loaded. This causes user confusion - "Which version is running, and if I can log in why to different parts of the system fail to show data?"

The fix is, for a new image install, stop any prior images (except for nginx, which provides the status updates).

(cherry picked from commit 40f021e)