OM-33021: XL: Use projected topology instead of original topology to get entity types for interpreted actions.
We are currently using the original topology to get entity types for interpreted actions. The original topology doesn't contain any entities that are provisioned as part of running the market algorithm, so we're not able to interpret those actions properly, leading to IllegalArgumentExceptions.

We should use the projected topology instead - the projected topology should contain all entities that are referenced to by the actions produced by the market. As an added bonus, this should have marginally better performance when running an analysis on scoped topologies.

Testing Done:
- Full build
- Ran an "Add Workload" plan that required provisioning a new entity. Verified there are no illegal argument exceptions in the logs.

Bugs closed: OM-33021

Reviewed at https://rbcommons.com/s/VMTurbo/r/22545/