OM-32176 RB:22652 Overwrite existing plan orchestrator state when restoring from diags
Restoration from diags will always occur on an empty instance of XL, rather than one
with data that is worth keeping. However, there may be default data present that is
populated when the component is started. Since this default data may not always be the
same, we need to overwrite any data already present when restoring from diags.