OM-33010: XL Stitching/History: Properly set MEM_ALLOCATION and CPU_ALLOCATION from MEM and CPU respectively, return per-commodity-key stats.
The VMM probe (and, presumably, others) don't necessarily send CPU_ALLOCATION with the same keys as the CPU commodity sold by the host. There is a CPU_ALLOCATION commodity for every VDC that the host is involved in, whereas there is only one CPU commodity. The same goes for MEM_ALLOCATION and MEMORY.

Solution: Adding a new method to OverprovisionCapacityPostStitchingOperation to allow sub-operations to ignore commodity keys when matching source to overprovision commodities. The *_ALLOCATION commodity operations set this to false, others set this to true.

Also, in StatsHistoryService we didn't create a new stat snapshot for every commodity name - commodity key - relation pair. Solution - use the commodity key (in addition to name and pair) when organizing stat records retrieved from the database.

Testing Done:
- Tested end-to-end with the VMM probe. Screenshots attached - CPU and memory allocation looks like it does in opsmgr.
- Full build, with additional unit tests.

Reviewed By: mark.laff, emanuele.maccherani
https://rbcommons.com/s/VMTurbo/r/22672/