OM-32629: XL API: Processing "related type" argument for scoped per-entity stats queries.
The XL API does not currently handle the "relatedType" argument when getting per-entity stats. The most noticeable effect of this is that the optimized improvements widget in plan results has no data when the plan scope is NOT a group of PMs (or cluster, or global environment).

The fix is to run a supply chain query when we have both specific scopes and a related type.

Testing Done:
- Build, with new unit test.
- Ran plan scoped to VM group - optimized widget shows hosts.
- Ran plan scoped to PM group, cluster, and global environment - optimized widget shows hosts as before.
- Looked at optimized improvements widget off nodes in the supply chain - shows the entities in the node (as in opsmgr)

Bugs closed: OM-32629

Reviewed at https://rbcommons.com/s/VMTurbo/r/22732/