OM-33816: XL Api/Repo: Returning SupplyChainNode members grouped by state.
The supply chain in develop doesn't render because we need to provide a new parameter - stateSummary - which is the number of entities by a particular state. The state summary is required in all supply chain responses, regardless of health summary and detail type.

Main change:
Adding a map from entity state (represented by an int) to the OIDs of entities of that state. Deprecating the member_oids field, and removing all uses of it, to reduce the duplication of information, since the member_oids should be calculatable from the oids-by-state map. This affects everything:
- The arangodb query for single-source supply chains.
- The supply chain calculation in SupplyChainSubgraph.
- The merging of single-source supply chains.
- The global supply chain query + conversion.

Minor changes:
- Removing unused old code to retrieve supply chain.
- Adding a minor optimization to the single-source suply chain query to reduce the amount of intermediate documents created.

Performance Impact:
- Global supply chain query is slower (up to 10% on my machine with the 100k test, although results are very all over the place)
- Single-source supply chain query is faster than before, but slower than just applying the optimization without returning the states.
- The difference seems largely due to returning more data from arango to repository.

Testing Done:
- Built api component and repository component.
- Started up XL (off develop branch), verified that supply chain and supply chain navigation work.

Bugs closed: OM-33816

Reviewed at https://rbcommons.com/s/VMTurbo/r/22908/