OM-33057: XL Proto: Pagination Parameters support.
Adding "PaginationParameters" and "PaginationResponse" protobuf messages to XL. These will be derived from the various subclasses of PaginationRequest, and included in the request/response messages of gRPC services that support pagination.

I'm opting to have a single message represent all the various subclasses (with a oneof field for subclass-specific parameters) to make it easier to keep pagination in sync between Java and protobuf. The total number of paginated services is fairly low, so this shouldn't be a problem for the forseeable future.

Testing Done:
- Compilation.

Bugs closed: OM-33057

Reviewed at https://rbcommons.com/s/VMTurbo/r/22890/