OM-34022: Plan execution never completes RB-23027: Roman
Account for a change in "startTime" convention for /stats/{plan-id}. It used to be that "no startTime" was used to trigger reading from the projected topology. Now the request comes with a 'startTime' in the future.

We changed the logic in RepositoryRpcService.getPlanTopologyStats to reflect the new convention on 'startTime'.

Further, the "hang" was due to a bug never calling 'responseObserver.onCompleted()' in the other case (plan source topology), which is not used by the UX currently and not implemented in XL at all.