OM-33677: implement commodity filtering in the History Component averaged entity stats call
RB-23001: David

Pass the 'filters' from the caller down to the methat that creates the select query. The only filter used in the UI for now is ```"type": "relation", "value": "sold/bought"```. I checked with UX team and that is all for now.

Also cleaned out some dead legacy code for handling the RelationType enum.