OM-33592: vmtdb migration performance degradation RB-23212: Emanuele
In 7.0.b12 there were some tests that only ran the 6.x DB migrations, not the XL additional migrations. The partitioning steps were extremely expensive, and so when we created a 7.x baseline the time for a database migration as part of these tests increased (e.g. from 6+ secs to 51 secs on a dev laptop).

In the new baseline we left in the partition creation statements added to the DB dump. These partition creation statements are unnecessary as the partitions will be recreated once the migration is complete, and the partition creation takes signficant time.

In this change we remove the initial partition creation, and so DB migration goes back down to ~10 secs

Also there was an old migration file (unused) which we deleted.