OM-34189 RB:23493 Create unified global disk-type-specific IOPS capacity settings
When calculating Storage Access capacity based on capacities for each type of disk,
use values configured globally in Consul rather than having individual settings for
each storage entity. If disk counts are present and Storage Access capacity is not
determined in one of the ways that overrides the calculation, calculate the capacity
by using the disk counts and flags within the DiskData property string. The count of
each disk type should be multiplied by that disk type's configured IOPS capacity,
and the products of each disk type should be added together. If certain flags are
present, the sum should then be multiplied by a configured factor associated with
the flag.