OM-35210 - Set autoCommit flag to true in DB connection() call. RB:23647
We use the connection() call for the select query from the BaseDbIO library
which is supposed to return an AutoCommitting connection. Since we use pooled
connections, for the same connection, we sometimes disable the autoCommiting
transaction(for doing an explicit transaction). But the flag is not reset to
true when the next time the connection() call is made. This was leading to
long running transaction holding table locks and preventing the vmtdb stats
table REPARTITIONs. The fix is to set the autoCommit to true every time
the connection() call is made.