OM-35316: Force computation of complete quote for current supplier RB:23870
We decide to move from a current provider to new provider when:
quote(new provider) < quote(current provider) * quote_factor
where quote_factor is a constant from market configuration with a default of 0.75.

As an optimization in the code that computes a quote that will early-exit the computation of the quote if we exceed the best quote found so far (because we know this supplier is not the best possible one).

Consider the following example: Suppose we are computing the quote for the current supplier and the best quote computed so far is 0.9. There are 3 commodities in the shopping list that add the following amounts to the overall quote sum [0.5, 0.5, 100.0]. If we early-exit as described above, after we add the cost of the second commodity to the quote sum, we will return a value of 1.0 for the quote of the current supplier.

As a result, when deciding whether to move to the supplier with the quote of 0.9, we will check 0.9 < 1.0 * 0.75 and decide not to move. However, if we used the correct complete quote value for the current supplier of 101.0, we would choose to take the move.

Expected impact:
* We will probably take a very small number of additional actions because the incorrect use of the early-exit quote computation resulted in us sometimes not taking moves we should take as in the example above.

Testing Done:
* Added new unit tests for QuoteFunctionFactory.
* Ran updated market against Verifone topology and saw sensible actions.