OM-35976: XL History: Add in-memory pagination for historical stats.
Before this change the history component ignores pagination parameters, and returns all entity stat records.

After this change, the history component will respect pagination parameters, and return only the next page of results. Right now I'm implementing this via a simple EntityStatsPaginator class, which does an in-memory sort of the records we get from the database. The next step (35980) is to do the pagination in the database.

Also adding a performance test for getting the first stage of entity stats, so that this in-memory solution can be a baseline.

Testing Done:
- Ran the HistoryPerformanceTest successfully.
- Unit tests for EntityStatsPaginator.

Reviewed By: emanuele_maccherani
https://rbcommons.com/s/VMTurbo/r/24330/